VENTURA: AppleScript change Siri Voice/ Lanugauge
MAC OS 13 VENTURA: 
This script lets you change language on Ventura, so that #you can get the Siri voice on multiple languages. 
Open automator -> Add run apple script -> Past the script #and create an app.

The script works by activating Voice over to navigate, so #you have to deactivate the VoiceOver welcome screen for it #to work correctly. On the first run you will be told that #the app lacks permission to keystroke. 
#What you have to do is go to system settings, Privacy & #Security -> Accessibility -> and turn on the app you #created. Use the pluss sign and add it if it did not appear automatically.